+++ GEOS_ChemGridComp.F90
@@ -650,6 +650,18 @@ contains
 
   END IF
 
+! GOCART2G <=> GMICHEM coupling ...
+! ---------------------------------
+  IF(myState%enable_GMICHEM .AND. TRIM(providerName) == "GOCART2G") THEN
+
+   IF(myState%enable_GOCART2G) &
+
+    CALL MAPL_AddConnectivity ( GC, &
+         SHORT_NAME  = (/'AERO'/),                &
+         DST_ID = GMICHEM, SRC_ID = GOCART2G, __RC__  )
+
+  END IF
+
 ! GOCART.data <=> GMICHEM coupling ...
 ! ------------------------------------
   IF(myState%enable_GMICHEM .AND. TRIM(providerName) == "GOCART.data") THEN
@@ -772,6 +784,15 @@ contains
    ENDIF
   ENDIF
 
+! GEOS-Chem import of CO2 
+! -----------------------------
+  IF(myState%enable_GEOSCHEM .AND. myState%enable_GOCART .AND. chemReg%doing_CO2) then
+   CALL MAPL_AddConnectivity ( GC, &
+       SRC_NAME  = (/"GOCART::CO2"/), &
+       DST_NAME  = (/"GOCART_CO2"/), &
+       DST_ID=GEOSCHEM, SRC_ID=GOCART, __RC__  )
+  ENDIF
+
 ! HEMCO connections to CHEMENV
 ! -----------------------------
   ! Default values:
@@ -857,8 +878,8 @@ contains
      END IF
 
      ! make sure we don't have inconsistent MEGAN flags
-     IF ( doMEGANviaHEMCO .eqv. .TRUE.    .AND.  &
-          doMEGANemission .eqv. .FALSE. ) THEN
+     IF ( ( doMEGANviaHEMCO .eqv. .TRUE. )    .AND.  &
+          ( doMEGANemission .eqv. .FALSE. ) ) THEN
         PRINT*,'Inconsistent GMI flags: doMEGANviaHEMCO==T, doMEGANemission==F'
         STATUS=99
         VERIFY_(STATUS)